Move App to src dir
